+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.3] - 2026-06-28
+
 ### Changed
 
 - Interface-error alerting now skips bridge, lagg and tunnel pseudo-interfaces